Revert "mrd-1970_ppcs-user-search-screen: search screen for PPCS users journey (#952)"
This reverts commit 50b9f22f54c94ed52a8fecb598e78255f62acd39.
