+++ server/views/pages/personSearchResults.njk
@@ -1,0 +1,44 @@
+{% extends "../partials/layout.njk" %}
+
+
+{% set pageTitle = makePageTitle({ pageHeading: "Search results"}) %}
+
+{% block content %}
+    <h1 class="govuk-heading-xl">Search results</h1>
+    <p class="govuk-body-l">
+        <span class='govuk-!-padding-right-2 govuk-!-font-weight-bold'>CRN: {{ crn }}</span>
+        <a href="/search-by-crn" class="govuk-link">Change <span class="govuk-visually-hidden">search term</span></a>
+    </p>
+
+    {% if persons.length %}
+    <table class="govuk-table">
+        <caption class="govuk-table__caption govuk-visually-hidden">Persons found</caption>
+        <thead class="govuk-table__head">
+        <tr class="govuk-table__row">
+            <th scope="col" class="govuk-table__header">Name</th>
+            <th scope="col" class="govuk-table__header">CRN</th>
+            <th scope="col" class="govuk-table__header">Date of birth</th>
+        </tr>
+        </thead>
+        <tbody class="govuk-table__body">
+        {% for person in persons %}
+            <tr class="govuk-table__row" data-qa='row-{{ person.crn }}'>
+                <td class="govuk-table__cell">
+                    {% if person.userExcluded %}
+                        You are excluded from viewing this case
+                    {% elseif person.userRestricted %}
+                        This is a restricted case
+                    {% endif %}
+                    <a href='/cases/{{ person.crn }}/overview' data-qa='name'>{% if person.userExcluded or person.userRestricted %}(read more information){% else %}{{ person.name }}{% endif %}</a>
+                </td>
+                <td class="govuk-table__cell">{{ person.crn }}</td>
+                <td class="govuk-table__cell">{{ formatDateTimeFromIsoString({isoDate: person.dateOfBirth}) }}</td>
+            </tr>
+        {% endfor %}
+        </tbody>
+    </table>
+    {% else %}
+    <p class='govuk-body'>No results</p>
+    {% endif %}
+
+{% endblock %}
